test(e2e): remove legacy tests with new version.
